[修正] #1823 スケジューラーのテストを php-cs-fixer / phpstan の規約に合わせた
CLAUDE.md の「新規・変更したファイルは php-cs-fixer と phpstan（level 9）を通す」を
満たしていなかった。既にコミット済みのテスト（#1833）は level 9 で無警告なので、
同じ状態に揃える。phpstan の指摘は 293 件から 0 件になった。

- phpstan.neon の scanFiles にスケジューラ関連のクラスを追加した。テスト側は実行時に
  require するが、パスを変数で組み立てるため phpstan が追えない。FRMigrationClass.php と
  同じ扱いにする
- テストが参照するクラスの読み込みをファイルスコープへ移した。既存テストと同じ書き方
- 列の値を取り出す箇所を型付きの getColString() / getColInt() に置き換えた。
  mixed のキャストが level 9 で通らないため
- mktime() の戻り値（int|false）、mysqli::query() の戻り値（mysqli_result|true）、
  $GLOBALS['dbconfig']（mixed）の扱いを明示した
- リフレクション経由の呼び出しに型付きの受け取りを用意した

あわせて、テストが実際に検証している振る舞いと食い違っていた 2 か所の
docblock を実態に合わせた。

- FR_CronDispatcher::tailLogFile() の $file は null を受け付ける（未指定なら空を返す）
- Vtiger_Cron::computeNextWeeklyRunAt() の $weekdays はカンマ区切りの文字列も受け付ける
  （説明文には書かれていたが @param の型に無かった）

+++ tests/Unit/Cron/FixedTimeScheduleTest.php
@@ -32,7 +32,7 @@ final class FixedTimeScheduleTest extends TestCase
     /** 基準時刻: 2026-08-25 10:07:33（火曜日） */
     private function reference(): int
     {
-        return mktime(10, 7, 33, 8, 25, 2026);
+        return (int) mktime(10, 7, 33, 8, 25, 2026);
     }
 
     // ------------------------------------------------------------ 毎日
@@ -40,7 +40,7 @@ final class FixedTimeScheduleTest extends TestCase
     public function test_I1_指定時刻がまだ来ていなければ当日になる(): void
     {
         self::assertSame(
-            mktime(15, 30, 0, 8, 25, 2026),
+            (int) mktime(15, 30, 0, 8, 25, 2026),
             Vtiger_Cron::computeNextDailyRunAt(15 * 60 + 30, $this->reference()),
             'I1 15:30 指定・10:07 時点 → 当日 15:30'
         );
@@ -49,13 +49,13 @@ final class FixedTimeScheduleTest extends TestCase
     public function test_I2_指定時刻を過ぎていれば翌日になる(): void
     {
         self::assertSame(
-            mktime(3, 0, 0, 8, 26, 2026),
+            (int) mktime(3, 0, 0, 8, 26, 2026),
             Vtiger_Cron::computeNextDailyRunAt(3 * 60, $this->reference()),
             'I2 03:00 指定・10:07 時点 → 翌日 03:00'
         );
         self::assertSame(
-            mktime(10, 7, 0, 8, 26, 2026),
-            Vtiger_Cron::computeNextDailyRunAt(10 * 60 + 7, mktime(10, 7, 0, 8, 25, 2026)),
+            (int) mktime(10, 7, 0, 8, 26, 2026),
+            Vtiger_Cron::computeNextDailyRunAt(10 * 60 + 7, (int) mktime(10, 7, 0, 8, 25, 2026)),
             'I2 境界: ちょうど指定時刻 → 翌日'
         );
     }
@@ -63,18 +63,18 @@ final class FixedTimeScheduleTest extends TestCase
     public function test_I3_境界値_0時_23時59分_月末をまたぐ場合(): void
     {
         self::assertSame(
-            mktime(0, 0, 0, 8, 26, 2026),
+            (int) mktime(0, 0, 0, 8, 26, 2026),
             Vtiger_Cron::computeNextDailyRunAt(0, $this->reference()),
             'I3 00:00 指定 → 翌日 00:00'
         );
         self::assertSame(
-            mktime(23, 59, 0, 8, 25, 2026),
+            (int) mktime(23, 59, 0, 8, 25, 2026),
             Vtiger_Cron::computeNextDailyRunAt(1439, $this->reference()),
             'I3 23:59 指定 → 当日 23:59'
         );
         self::assertSame(
-            mktime(1, 0, 0, 9, 1, 2026),
-            Vtiger_Cron::computeNextDailyRunAt(60, mktime(23, 0, 0, 8, 31, 2026)),
+            (int) mktime(1, 0, 0, 9, 1, 2026),
+            Vtiger_Cron::computeNextDailyRunAt(60, (int) mktime(23, 0, 0, 8, 31, 2026)),
             'I3 月末をまたぐ（8/31 23:00 時点で 01:00 指定）→ 9/1 01:00'
         );
     }
@@ -82,12 +82,12 @@ final class FixedTimeScheduleTest extends TestCase
     public function test_I4_範囲外の時刻は0時として扱う(): void
     {
         self::assertSame(
-            mktime(0, 0, 0, 8, 26, 2026),
+            (int) mktime(0, 0, 0, 8, 26, 2026),
             Vtiger_Cron::computeNextDailyRunAt(1440, $this->reference()),
             'I4 範囲外（1440）は 0 時として扱う'
         );
         self::assertSame(
-            mktime(0, 0, 0, 8, 26, 2026),
+            (int) mktime(0, 0, 0, 8, 26, 2026),
             Vtiger_Cron::computeNextDailyRunAt(-10, $this->reference()),
             'I4 範囲外（負）は 0 時として扱う'
         );
@@ -100,22 +100,22 @@ final class FixedTimeScheduleTest extends TestCase
         $reference = $this->reference();
 
         self::assertSame(
-            mktime(9, 0, 0, 8, 28, 2026),
+            (int) mktime(9, 0, 0, 8, 28, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt(5, 9 * 60, $reference),
             'I8 火曜 10:07 時点で金曜 09:00 指定 → 同じ週の金曜'
         );
         self::assertSame(
-            mktime(9, 0, 0, 8, 31, 2026),
+            (int) mktime(9, 0, 0, 8, 31, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt(1, 9 * 60, $reference),
             'I8 火曜 10:07 時点で月曜 09:00 指定 → 翌週の月曜'
         );
         self::assertSame(
-            mktime(15, 0, 0, 8, 25, 2026),
+            (int) mktime(15, 0, 0, 8, 25, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt(2, 15 * 60, $reference),
             'I8 境界: 当日（火曜）で時刻前 → 当日'
         );
         self::assertSame(
-            mktime(9, 0, 0, 9, 1, 2026),
+            (int) mktime(9, 0, 0, 9, 1, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt(2, 9 * 60, $reference),
             'I8 境界: 当日（火曜）で時刻後 → 翌週'
         );
@@ -137,27 +137,27 @@ final class FixedTimeScheduleTest extends TestCase
         $reference = $this->reference();
 
         self::assertSame(
-            mktime(9, 0, 0, 8, 26, 2026),
+            (int) mktime(9, 0, 0, 8, 26, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt([1, 3, 5], 9 * 60, $reference),
             'I15 月・水・金指定・火曜 10:07 時点 → 直近の水曜'
         );
         self::assertSame(
-            mktime(9, 0, 0, 8, 26, 2026),
+            (int) mktime(9, 0, 0, 8, 26, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt('1,3,5', 9 * 60, $reference),
             'I15 カンマ区切りの文字列でも同じ結果になる'
         );
         self::assertSame(
-            mktime(15, 0, 0, 8, 25, 2026),
+            (int) mktime(15, 0, 0, 8, 25, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt([2, 4], 15 * 60, $reference),
             'I15 当日（火曜）を含み時刻前なら当日'
         );
         self::assertSame(
-            mktime(9, 0, 0, 8, 27, 2026),
+            (int) mktime(9, 0, 0, 8, 27, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt([2, 4], 9 * 60, $reference),
             'I15 当日（火曜）を含むが時刻後なら次の指定曜日（木曜）'
         );
         self::assertSame(
-            mktime(9, 0, 0, 8, 26, 2026),
+            (int) mktime(9, 0, 0, 8, 26, 2026),
             Vtiger_Cron::computeNextWeeklyRunAt([0, 1, 2, 3, 4, 5, 6], 9 * 60, $reference),
             'I15 全曜日を指定すると翌日になる（毎日と同じ間隔）'
         );
@@ -180,17 +180,17 @@ final class FixedTimeScheduleTest extends TestCase
         $reference = $this->reference();
 
         self::assertSame(
-            mktime(9, 0, 0, 8, 28, 2026),
+            (int) mktime(9, 0, 0, 8, 28, 2026),
             Vtiger_Cron::computeNextMonthlyRunAt(28, 9 * 60, $reference),
             'I10 8/25 10:07 時点で 28 日 09:00 指定 → 当月 8/28'
         );
         self::assertSame(
-            mktime(9, 0, 0, 9, 10, 2026),
+            (int) mktime(9, 0, 0, 9, 10, 2026),
             Vtiger_Cron::computeNextMonthlyRunAt(10, 9 * 60, $reference),
             'I10 8/25 10:07 時点で 10 日 09:00 指定 → 翌月 9/10'
         );
         self::assertSame(
-            mktime(9, 0, 0, 9, 25, 2026),
+            (int) mktime(9, 0, 0, 9, 25, 2026),
             Vtiger_Cron::computeNextMonthlyRunAt(25, 9 * 60, $reference),
             'I10 境界: 当日で時刻後 → 翌月'
         );
@@ -199,7 +199,7 @@ final class FixedTimeScheduleTest extends TestCase
     public function test_I11_月末指定は月の長さに追従する(): void
     {
         self::assertSame(
-            mktime(9, 0, 0, 8, 31, 2026),
+            (int) mktime(9, 0, 0, 8, 31, 2026),
             Vtiger_Cron::computeNextMonthlyRunAt(
                 Vtiger_Cron::RUN_ON_LAST_DAY_OF_MONTH,
                 9 * 60,
@@ -208,20 +208,20 @@ final class FixedTimeScheduleTest extends TestCase
             'I11 月末指定（8 月）→ 8/31'
         );
         self::assertSame(
-            mktime(9, 0, 0, 2, 28, 2027),
+            (int) mktime(9, 0, 0, 2, 28, 2027),
             Vtiger_Cron::computeNextMonthlyRunAt(
                 Vtiger_Cron::RUN_ON_LAST_DAY_OF_MONTH,
                 9 * 60,
-                mktime(10, 0, 0, 2, 1, 2027)
+                (int) mktime(10, 0, 0, 2, 1, 2027)
             ),
             'I11 月末指定（2027 年 2 月・平年）→ 2/28'
         );
         self::assertSame(
-            mktime(9, 0, 0, 2, 29, 2028),
+            (int) mktime(9, 0, 0, 2, 29, 2028),
             Vtiger_Cron::computeNextMonthlyRunAt(
                 Vtiger_Cron::RUN_ON_LAST_DAY_OF_MONTH,
                 9 * 60,
-                mktime(10, 0, 0, 2, 1, 2028)
+                (int) mktime(10, 0, 0, 2, 1, 2028)
             ),
             'I11 月末指定（2028 年 2 月・うるう年）→ 2/29'
         );
@@ -230,8 +230,8 @@ final class FixedTimeScheduleTest extends TestCase
     public function test_I12_その月に無い日を指定したら末日になる(): void
     {
         self::assertSame(
-            mktime(9, 0, 0, 2, 28, 2027),
-            Vtiger_Cron::computeNextMonthlyRunAt(31, 9 * 60, mktime(10, 0, 0, 2, 1, 2027)),
+            (int) mktime(9, 0, 0, 2, 28, 2027),
+            Vtiger_Cron::computeNextMonthlyRunAt(31, 9 * 60, (int) mktime(10, 0, 0, 2, 1, 2027)),
             'I12 2 月に 31 日指定 → その月の末日（2/28）'
         );
     }
@@ -239,8 +239,8 @@ final class FixedTimeScheduleTest extends TestCase
     public function test_I13_年をまたぐ場合(): void
     {
         self::assertSame(
-            mktime(9, 0, 0, 1, 5, 2027),
-            Vtiger_Cron::computeNextMonthlyRunAt(5, 9 * 60, mktime(10, 0, 0, 12, 20, 2026)),
+            (int) mktime(9, 0, 0, 1, 5, 2027),
+            Vtiger_Cron::computeNextMonthlyRunAt(5, 9 * 60, (int) mktime(10, 0, 0, 12, 20, 2026)),
             'I13 12/20 時点で 5 日指定 → 翌年 1/5'
         );
     }
@@ -255,19 +255,19 @@ final class FixedTimeScheduleTest extends TestCase
     {
         // 10:07:33 時点で 15 分周期 → 10:15:00
         self::assertSame(
-            mktime(10, 15, 0, 8, 25, 2026),
+            (int) mktime(10, 15, 0, 8, 25, 2026),
             Vtiger_Cron::computeNextRunAt(900, $this->reference()),
             '15 分周期は :00/:15/:30/:45 に揃う'
         );
         // 1 分周期 → 10:08:00
         self::assertSame(
-            mktime(10, 8, 0, 8, 25, 2026),
+            (int) mktime(10, 8, 0, 8, 25, 2026),
             Vtiger_Cron::computeNextRunAt(60, $this->reference()),
             '1 分周期は毎分 0 秒に揃う'
         );
         // 12 時間周期 → 12:00:00
         self::assertSame(
-            mktime(12, 0, 0, 8, 25, 2026),
+            (int) mktime(12, 0, 0, 8, 25, 2026),
             Vtiger_Cron::computeNextRunAt(43200, $this->reference()),
             '12 時間周期は 0:00 / 12:00 に揃う'
         );
@@ -276,9 +276,9 @@ final class FixedTimeScheduleTest extends TestCase
     public function test_遅れてもグリッドへ復帰し遅れを持ち越さない(): void
     {
         // 19:58 の回を取りこぼし、19:59:30 に完了した場合でも次回は 20:00:00
-        $finished = mktime(19, 59, 30, 8, 25, 2026);
+        $finished = (int) mktime(19, 59, 30, 8, 25, 2026);
         self::assertSame(
-            mktime(20, 0, 0, 8, 25, 2026),
+            (int) mktime(20, 0, 0, 8, 25, 2026),
             Vtiger_Cron::computeNextRunAt(60, $finished),
             '遅れて完了しても次のグリッドに戻る（完了時刻 + 周期にならない）'
         );
